Added files relating to game prediction
Added betting.R and associated CSV files.

Betting.R estimates the likelihood of the home team winning a game based on the relative strength of the two teams and whether either team is in the second game of a back-to-back. It compares this likelihood to betting odds (manually pulled from BCLC). The model then selects a team to bet on and estimates a risk-adjusted return.